requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0

# Web framework
streamlit>=1.28.0

# Data visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0
